+++ ai-logic/firebase-ai/src/testUtil/java/com/google/firebase/ai/JavaCompileTests.java
@@ -34,6 +34,7 @@ import com.google.firebase.ai.java.LiveModelFutures;
 import com.google.firebase.ai.java.LiveSessionFutures;
 import com.google.firebase.ai.java.TemplateChatFutures;
 import com.google.firebase.ai.java.TemplateGenerativeModelFutures;
+import com.google.firebase.ai.type.AspectRatio;
 import com.google.firebase.ai.type.BlockReason;
 import com.google.firebase.ai.type.Candidate;
 import com.google.firebase.ai.type.Citation;
@@ -51,7 +52,9 @@ import com.google.firebase.ai.type.GenerationConfig;
 import com.google.firebase.ai.type.HarmCategory;
 import com.google.firebase.ai.type.HarmProbability;
 import com.google.firebase.ai.type.HarmSeverity;
+import com.google.firebase.ai.type.ImageConfig;
 import com.google.firebase.ai.type.ImagePart;
+import com.google.firebase.ai.type.ImageSize;
 import com.google.firebase.ai.type.ImagenBackgroundMask;
 import com.google.firebase.ai.type.ImagenEditMode;
 import com.google.firebase.ai.type.ImagenEditingConfig;
@@ -144,6 +147,11 @@ public class JavaCompileTests {
         .setResponseMimeType("image/jxl")
         .setResponseModalities(List.of(ResponseModality.TEXT, ResponseModality.TEXT))
         .setResponseSchema(getSchema())
+        .setImageConfig(
+            ImageConfig.builder()
+                .setAspectRatio(AspectRatio.LANDSCAPE_21x9)
+                .setImageSize(ImageSize.SIZE_512)
+                .build())
         .build();
   }